Homepage como método das classes,
MuseuRobertoMarino.py teve o nome trocado para main.py
MuseuDino.py teve o nome trocado para funcoes.py
As funções homepage foram colocadas como métodos acessar_homepage() em Pessoa (para polimorfismo), Funcionario e Visitante.
Pequenas correções no código em coisas que estavam dando erro.
Coloquei nome e dataNascimento como atributos privados.
